Move failpoint support code to utils.
To enable them in safekeeper as well.
